requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.44.2
datasets==2.14.2
evaluate==0.4.6
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.13.2
geopandas==1.1.1
shapely==2.1.2
accelerate==0.33.0
python-dotenv==1.0.0
